Przeniesienie rzeczy labiryntu do labirynt*
